검색 옵션 전체인 경우 오류 수정, 댓글 있는 게시물 delete조작 에러처리

+++ src/main/java/com/naver/jihyunboard/board/service/BoardService.java
@@ -111,9 +111,15 @@ public class BoardService {
 	}
 
 	@Transactional
-	public void deleteBoard(int boardNum) throws Exception {
-		boardRepository.deleteFile(boardNum);
-		boardRepository.deleteBoard(boardNum);
+	public String deleteBoard(int boardNum) throws Exception {
+		int replyCount = boardRepository.getReplyCount(boardNum);
+		if (replyCount > 0) {
+			return "no";
+		} else {
+			boardRepository.deleteFile(boardNum);
+			boardRepository.deleteBoard(boardNum);
+			return "ok";
+		}
 	}
 
 	public List<UploadFile> getFileList(int bno) {
